develop(1219-dynamodb-resolve-host-exception): fix secretsmanager entities
* fix(1219-dynamodb-resolve-host-exception): fix resolve exception, fix…

* fix(1219-dynamodb-resolve-host-exception): fix resolve exception, fix…

* fix(1219-dynamodb-resolve-host-exception): fix resolve exception, fix…

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* fix(1219-dynamodb-resolve-host-exception): add scan-item.sh

* fix(1219-dynamodb-resolve-host-exception): add scan-item.sh

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager RES…

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager RES…

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager ent…

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

+++ src/service/src/secretsmanager/SecretsManagerService.cpp
@@ -126,7 +126,7 @@ namespace AwsMock::Service {
             response.versionId = secret.versionId;
             if (!secret.secretString.empty()) {
                 const std::string base64Decoded = Core::Crypto::Base64Decode(secret.secretString);
-                int len = base64Decoded.length();
+                int len = static_cast<int>(base64Decoded.length());
                 response.secretString = std::string(reinterpret_cast<char *>(Core::Crypto::Aes256DecryptString((unsigned char *) base64Decoded.c_str(), &len, (unsigned char *) _kmsKey.c_str())));
             } /*else if (!secret.secretString.empty()) {
                 std::string base64Decoded = Core::Crypto::Base64Decode(secret.secretString);
